Multi-pet support, proc parity & responsive UI
Introduce THJ-style multi-pet persistence and runtime behavior, proc parity controls, and a responsive server manager layout. Key changes: save/load multiple permanent pets and legacy suspended-slot migration (new SQL migration); add PetInfo vector and related pet handling (focused pet, pet commands, limits via Custom:AbsolutePetLimit, suspended pet accessor); align proc behavior (PetProcRateCap, ExcludeTempPetsFromProcChanceSPA, MultipleTwoHandedProcs, ProcReflectPercentage) and ranged double/triple-attack/skill checks (DoubleAttackSkillRanged); add attack mode/weapon appearance and autoskill timing; various safety casts and minor fixes. Also add layout_mode to server_manager with auto/portrait/landscape reflow and a layout selector, and update ruletypes/docs to reflect added/removed rules and new multiclass/mult-pet config options.

+++ zone/mob.cpp
@@ -599,7 +599,7 @@ Mob::~Mob()
 			GetPet()->BuffFadeByEffect(SpellEffect::Charm);
 		}
 		else {
-			SetPet(0);
+			SetPet(static_cast<uint16>(0));
 		}
 	}
 
@@ -8348,7 +8348,7 @@ void Mob::SetFeigned(bool in_feigned) {
 	if (in_feigned)	{
 		if (IsClient()) {
 			if (RuleB(Character, FeignKillsPet)){
-				SetPet(0);
+				SetPet(static_cast<uint16>(0));
 			}
 			CastToClient()->SetHorseId(0);
 		}
